requirements.txt
# === Core App & Interface ===
streamlit
streamlit-chat
python-dotenv
watchdog

# === OpenAI + LLMs ===
openai
tiktoken

# === File Parsing (PDF, Word, Excel, etc.) ===
PyPDF2
PyMuPDF
pypdf
pdfminer.six
python-docx
openpyxl
pandas
xlrd
csvkit

# === Google Drive & Auth ===
google-api-python-client
google-auth
google-auth-oauthlib

# === Web Search ===
duckduckgo-search
requests
beautifulsoup4
lxml
httpx
urllib3

# === Utilities & Asynchronous Support ===
numpy
toolz
loguru
tenacity
aiohttp

# === Developer & Debugging ===
ipython
jupyter
black
pre-commit
